refactor: Piece 접근 지정자 수정
