+++ src/main/java/softeer2nd/chess/ChessView.java
@@ -1,7 +1,7 @@
 package softeer2nd.chess;
 
-import softeer2nd.chess.pieces.Piece;
-import softeer2nd.chess.pieces.enums.Type;
+import softeer2nd.chess.pieces.piece.Piece;
+import softeer2nd.chess.pieces.piece.Type;
 import softeer2nd.chess.util.StringUtils;
 
 import java.util.ArrayList;
