+++ src/main/java/softeer2nd/chess/pieces/Blank.java
@@ -1,8 +1,9 @@
 package softeer2nd.chess.pieces;
 
 import softeer2nd.chess.ChessGame;
-import softeer2nd.chess.pieces.enums.Color;
-import softeer2nd.chess.pieces.enums.Type;
+import softeer2nd.chess.pieces.piece.Color;
+import softeer2nd.chess.pieces.piece.Type;
+import softeer2nd.chess.pieces.piece.Piece;
 
 public class Blank extends Piece {
     private Blank(String position) {
