--- src/main/java/softeer2nd/chess/pieces/Direction.java
@@ -1,35 +1,0 @@
-package softeer2nd.chess.pieces;
-
-import java.util.Arrays;
-import java.util.List;
-
-public enum Direction {
-    NNE(1, 2),
-    NNW(-1, 2),
-    SSE(1, -2),
-    SSW(-1, -2),
-    EEN(2, 1),
-    EES(2, -1),
-    WWN(-2, 1),
-    WWS(-2, -1);
-
-    private int xDegree;
-    private int yDegree;
-
-    private Direction(int xDegree, int yDegree) {
-        this.xDegree = xDegree;
-        this.yDegree = yDegree;
-    }
-
-    public int getXDegree() {
-        return xDegree;
-    }
-
-    public int getYDegree() {
-        return yDegree;
-    }
-
-    public static List<Direction> knightDirection() {
-        return Arrays.asList(NNE, NNW, SSE, SSW, EEN, EES, WWN, WWS);
-    }
-}
