+++ src/main/java/softeer2nd/chess/pieces/King.java
@@ -1,11 +1,12 @@
 package softeer2nd.chess.pieces;
 
 import softeer2nd.chess.ChessGame;
-import softeer2nd.chess.pieces.enums.Color;
-import softeer2nd.chess.pieces.enums.Type;
+import softeer2nd.chess.pieces.piece.Color;
+import softeer2nd.chess.pieces.piece.Type;
 import softeer2nd.chess.pieces.exception.InvalidMoveException;
+import softeer2nd.chess.pieces.piece.Piece;
 
-public class King extends Piece{
+public class King extends Piece {
     private King(Color color, String position) {
         super(color, Type.KING, position);
     }
