+++ src/main/java/softeer2nd/chess/pieces/Knight.java
@@ -1,13 +1,13 @@
 package softeer2nd.chess.pieces;
 
 import softeer2nd.chess.ChessGame;
-import softeer2nd.chess.pieces.enums.Color;
-import softeer2nd.chess.pieces.enums.Type;
+import softeer2nd.chess.pieces.enums.Direction;
+import softeer2nd.chess.pieces.piece.Color;
+import softeer2nd.chess.pieces.piece.Type;
 import softeer2nd.chess.pieces.exception.InvalidMoveException;
+import softeer2nd.chess.pieces.piece.Piece;
 
-import java.util.Optional;
-
-public class Knight extends Piece{
+public class Knight extends Piece {
     private Knight(Color color, String position) {
         super(color, Type.KNIGHT, position);
     }
