+++ src/main/java/softeer2nd/chess/pieces/Queen.java
@@ -1,9 +1,10 @@
 package softeer2nd.chess.pieces;
 
 import softeer2nd.chess.ChessGame;
+import softeer2nd.chess.pieces.piece.Piece;
 import softeer2nd.chess.util.StringUtils;
-import softeer2nd.chess.pieces.enums.Color;
-import softeer2nd.chess.pieces.enums.Type;
+import softeer2nd.chess.pieces.piece.Color;
+import softeer2nd.chess.pieces.piece.Type;
 import softeer2nd.chess.pieces.exception.InvalidMoveException;
 
 public class Queen extends Piece {
