+++ src/main/java/softeer2nd/chess/pieces/enums/Direction.java
@@ -4,15 +4,6 @@ import java.util.Arrays;
 import java.util.List;
 
 public enum Direction {
-    NORTH(0, 1),
-    NORTHEAST(1, 1),
-    EAST(1, 0),
-    SOUTHEAST(1, -1),
-    SOUTH(0, -1),
-    SOUTHWEST(-1, -1),
-    WEST(-1, 0),
-    NORTHWEST(-1, 1),
-
     NNE(1, 2),
     NNW(-1, 2),
     SSE(1, -2),
@@ -38,27 +29,7 @@ public enum Direction {
         return yDegree;
     }
 
-    public static List<Direction> linearDirection() {
-        return Arrays.asList(NORTH, EAST, SOUTH, WEST);
-    }
-
-    public static List<Direction> diagonalDirection() {
-        return Arrays.asList(NORTHEAST, SOUTHEAST, SOUTHWEST, NORTHWEST);
-    }
-
-    public static List<Direction> everyDirection() {
-        return Arrays.asList(NORTH, EAST, SOUTH, WEST, NORTHEAST, SOUTHEAST, SOUTHWEST, NORTHWEST);
-    }
-
     public static List<Direction> knightDirection() {
         return Arrays.asList(NNE, NNW, SSE, SSW, EEN, EES, WWN, WWS);
-    }
-
-    public static List<Direction> whitePawnDirection() {
-        return Arrays.asList(NORTH, NORTHEAST, NORTHWEST);
-    }
-
-    public static List<Direction> blackPawnDirection() {
-        return Arrays.asList(SOUTH, SOUTHEAST, SOUTHWEST);
     }
 }
