+++ src/main/java/softeer2nd/chess/pieces/piece/Color.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess.pieces.enums;
+package softeer2nd.chess.pieces.piece;
 
 public enum Color {
     // 반복문 돌릴시 NO_COLOR 문제
