+++ src/main/java/softeer2nd/chess/pieces/piece/Type.java
@@ -1,7 +1,6 @@
-package softeer2nd.chess.pieces.enums;
+package softeer2nd.chess.pieces.piece;
 
 public enum Type {
-    // 반복문 돌릴시 NO_PIECE 문제
     PAWN('p', 1.0), ROOK('r', 5.0), KNIGHT('n', 2.5),
     BISHOP('b', 3.0), QUEEN('q', 9.0), KING('k', 0.0),
     NO_PIECE('.', 0.0);
